Remove old orchestrator path
