+++ .gitignore
@@ -1,0 +1,17 @@
+# Byte-compiled / cache
+__pycache__/
+*.py[cod]
+*.pyo
+
+# Virtual environments
+venv/
+.env
+
+# Logs and keys
+logs/
+keys/
+*.pem
+
+# OS files
+.DS_Store
+Thumbs.db
